feat: Phase 2E - Item detail view with navigation
Implement full-screen detail view for media items with:
- Hero banner with backdrop image, title, metadata badges
- Overview and detailed information sections
- "More Like This" recommendations carousel
- Deep navigation support (detail → detail → detail)
- Progress tracking display
- Play/Resume and Favorite action buttons

Technical implementation:
- ItemDetailViewModel for state management and API calls
- ItemDetailView with hero header, metadata sections
- NavigationCoordinator with NavigationPath for deep stacking
- ItemDetailViewWrapper with @StateObject for ViewModel lifecycle
- getSimilarItems API endpoint integration
- HomeViewModel extracted from HomeView for better separation

Includes bug fix for ImageCache Int64 type casting.

Known issue: ScrollView position after 3+ navigations (deferred)

+++ MediaMio/Services/ImageCache.swift
@@ -189,12 +189,12 @@ class ImageCache {
             print("💿 Disk cache size: \(totalSize / 1024 / 1024) MB")
 
             // If over limit, remove oldest files
-            if totalSize > maxDiskCacheSize {
+            if totalSize > Int64(maxDiskCacheSize) {
                 // Sort by date (oldest first)
                 files.sort { $0.date < $1.date }
 
                 var removedSize: Int64 = 0
-                let targetRemovalSize = totalSize - (maxDiskCacheSize * 8 / 10)  // Remove down to 80% of max
+                let targetRemovalSize = totalSize - (Int64(maxDiskCacheSize) * 8 / 10)  // Remove down to 80% of max
 
                 for file in files {
                     if removedSize >= targetRemovalSize {
